Update Prettier to 3.0.3
I am going to introduce a submodule and want to make use of a
.prettierignore file to tell Prettier to ignore it. However, we are
already overriding the --ignore-path flag to tell it to ignore the files
described by .gitignore.

Prettier 3 ignores the files described by .gitignore by default. This
means that we can now remove our usage of --ignore-path and it’ll ignore
the files described by .gitignore _and_ .prettierignore.